+++ fxn/cli/auth.py
@@ -15,20 +15,20 @@ app = Typer(no_args_is_help=True)
 @app.command(name="login", help="Login to Function.")
 def login (
     access_key: str=Argument(..., help="Function access key.", envvar="FXN_ACCESS_KEY")
-) -> None:
+):
     user = User.retrieve(access_key=access_key)
     user = asdict(user) if user else None
     _set_access_key(access_key if user is not None else None)
     print_json(data=user)
 
 @app.command(name="status", help="Get current authentication status.")
-def auth_status () -> None:
+def auth_status ():
     user = User.retrieve(access_key=get_access_key())
     user = asdict(user) if user else None
     print_json(data=user)
 
 @app.command(name="logout", help="Logout from Function.")
-def logout () -> None:
+def logout ():
     _set_access_key(None)
     print("Successfully logged out of Function")
 
@@ -46,6 +46,12 @@ def get_access_key () -> str:
         return f.read()
 
 def _set_access_key (key: str):
+    """
+    Set the CLI access key.
+
+    Parameters:
+        key (str); CLI access key.
+    """
     credentials_path = Path.home() / ".fxn" / "credentials"
     credentials_path.parent.mkdir(parents=True, exist_ok=True)
     if key: